Update DIVISION.C
Made it so that division shows value up to 2 digits

+++ DIVISION.C
@@ -8,7 +8,7 @@ int main()
 	scanf("%d", &num1);
 	printf("Enter the second number: ");
 	scanf("%d", &num2);
-	printf("%d divided by %d is: %d\n", num1, num2, num1/num2);
+	printf("%.0f divided by %d is: %.2f\n", (float)num1, num2, (float)num1/num2);
 
 	return 0;
 }
